Satisfy naked approval CI checks
